Removes all non-FinanceSession code. Adds example session.
All of the functionality that had existed in the messy, pre-FS
code is now improved and duplicated in FinanceSession, so I removed
it. An example session testing the current capabilities of FS replaces
the old, hacky, test session. Syntax errors, nonexistent variables,
and silly mistakes (passing names instead of data to printing funcs,
for example) have been fixed.